requirements.txt
scipy
numpy
pillow
sklearn

